Reduce serializing
